% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filmweb_charakterystyki.R
\name{filmweb_charakterystyki}
\alias{filmweb_charakterystyki}
\title{Pobierz charakterysyki filmow z Filmwebu}
\usage{
filmweb_charakterystyki(sciezka_do_tabeli_danych, gdzie_zapisac_tabele,
  gdzie_zapisac_obsade, poczatek = 1, koniec = nrow(tabela_danych),
  czy_dopisac_do_pliku = FALSE)
}
\arguments{
\item{sciezka_do_tabeli_danych}{sciezka_dostepu do tabeli zawierajacej kolumne o nazwie \code{link_filmweb} oraz \code{tytul_filmweb}}

\item{gdzie_zapisac_tabele}{sciezka dostepu do folderu, w ktorym ma byc zapisana ostateczna tabela ze wskaznikami}

\item{gdzie_zapisac_obsade}{sciezka dostepu do folderu, w ktorym maja byc zapisane pliki .txt z obsada poszczegolnych filmow (beda one zapisane w formie: uproszczonytytulfilmu_rokprodukcji_ocena_obsady.txt)}

\item{poczatek}{pierwszy wiersz tabeli, ktory chcemy analizowac}

\item{koniec}{ostatni wiersz tabeli, ktory chcemy analizowac (domyslnie wszystkie wiersze)}

\item{czy_dopisac_do_pliku}{jesli \code{TRUE} dane zostana dopisane, a nie nadpisane do istniejacego pliku (zostana wtedy zapisane bez nazw kolumn)}
}
\value{
Funkcja \code{filmy_z_filmwebu} zwraca ramke danych o 16 kolumnach (2 wejsciowe, jedna bedaca znacznikiem filmwebu i 13 roznych charakterysyk - box_office, dluzszy_opis_filmu, gatunek, ile_osob_chce_zobaczyc, kraj_produkcji, krotki_opis_filmu, ktory_w_world_rankingu, liczba_komentarzy, liczba_oddanych_glosow, minuty_trwania_filmu, ocena_filmu, premiera_polska, premiera_swiat) zapisana w pliku o rozszerzeniu .txt oraz tyle plikow .txt z obsada i jej ocena, ile bylo filmow w wejsciowej ramce danych, majacych link do Filmwebu.
}
\description{
Funkcja \code{filmweb_charakterystyki()} szuka roznych charakterystyk dla filmow z Filmwebu.
}

