% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filmweb_tytul_rok_link.R
\name{filmweb_tytul_rok_link}
\alias{filmweb_tytul_rok_link}
\title{Pobierz liste filmow z portalu Filmweb}
\usage{
filmweb_tytul_rok_link(poczatek = 1, koniec = NULL,
  sciezka_do_pliku_txt_gdzie_zapisac, czy_dopisac_do_pliku = FALSE)
}
\arguments{
\item{poczatek}{numer filmu, od ktorego chcemy zaczac pobieranie (domyslnie 1)}

\item{koniec}{numer filmu, na ktorym chcemy skonczyc pobieranie (domyslnie wszystkie dostepne)}

\item{sciezka_do_pliku_txt_gdzie_zapisac}{sciezka do pliku .txt, w ktorym chcemy zapisac nasza ramke danych (zapisywanie odbywa sie na biezaco - tak wiec w przypadku wystapienia bledu, wszystkie sciagniete do tej pory dane zostana zapisane)}

\item{czy_dopisac_do_pliku}{jesli \code{TRUE} dane zostana dopisane, a nie nadpisane do istniejacego pliku (zostana wtedy zapisane bez nazw kolumn)}
}
\value{
Funkcja \code{filmweb_tytul_rok_link} zwraca ramke danych o trzech kolumnach. Pierwsza kolumna to \code{tytul_filmweb} zawierajaca tytyly filmow znalezionych przez funkcje, druga to \code{rok_produkcji} z rokiem produkcji filmu, a trzecia o nazwie \code{link_filmweb} przechowuje linki do strony danego filmu na Filmwebie.
}
\description{
Funkcja \code{filmweb_tytul_rok_link()} zapisuje na dysku najczesciej komentowane filmy na Filmwebie.
}

