% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imdb_charakterystyki.R
\name{imdb_charakterystyki}
\alias{imdb_charakterystyki}
\title{Pobierz podstawowe informacje na temat wybranych filmow na portalu IMDb}
\usage{
imdb_charakterystyki(sciezka_wejscie, sciezka_wyjscie)
}
\arguments{
\item{sciezka_wejscie}{Plik tekstowy zawierajace nastepujace dane: tytuly filmow, lata produkcji, linki do stron filmow na portalu IMDb.}
}
\value{
Funkcja \code{imdb_charakterystyki} zwraca ramke danych zawierajaca nastepujace dane: tytul filmu, rok produkcji, link, polski tytul, czas trwania (w minutach), ocena (w skali od 1 do 10), liczba glosow, liczba recenzji uzytkownikow, liczba recenzji krytykow, krotkie streszczenie.
}
\description{
Funkcja \code{imdb_charakterystyki} znajduje podstawowe informacje na temat wybranych filmow znajdujacych sie w bazie portalu IMDb i zapisuje je w pliku tekstowym.
}

