% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imdb_tytul_rok_link.R
\name{imdb_tytul_rok_link}
\alias{imdb_tytul_rok_link}
\title{Pobierz liste filmow z portalu IMDb}
\usage{
imdb_tytul_rok_link(liczba_filmow = NULL, sciezka_wyjscie)
}
\arguments{
\item{liczba_filmow}{Liczba tytulow, ktore chcemy pobrac (w kolejnosci malejacej ze wzgledu na liczbe glosow).}

\item{sciezka_wyjscie}{Sciezka do pliku tekstowego, w ktorym chcemy zapisac pobrane dane.}
}
\value{
Funkcja \code{imdb_tytul_rok_link} zwraca ramke danych o trzech kolumnach. Pierwsza kolumna \code{tytul_imdb} zawiera tytuly pobranych filmow. Druga kolumna \code{rok_produkcji} zawiera informacje o latach powstania filmow. Trzecia kolumna \code{link_imdb} zawiera linki do stron filmow na portalu IMDb.
}
\description{
Funkcja \code{imdb_tytul_rok_link} zapisuje w pliku tekstowym tytuly najczesciej komentowanych filmow na portalu IMDb, wraz z latami produkcji i linkami do stron poszczegolnych filmow.
}

