% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polacz_tabele_z_filmami.R
\name{polacz_tabele_z_filmami}
\alias{polacz_tabele_z_filmami}
\title{Polacz tabele z filmami}
\usage{
polacz_tabele_z_filmami(sciezka_do_tabel, sciezka_do_pliku_txt_gdzie_zapisac)
}
\arguments{
\item{sciezka_do_tabel}{wektor napisow, bedacy sciezka dostepu do tabel, ktore chcemy polaczyc (kolumny tabel powinny miec nazwy: \code{tytul_portal}, \code{rok_produkcji}, gdzie za \code{portal} trzeba wstawic odpowiednia nazwe portalu).}

\item{sciezka_do_pliku_txt_gdzie_zapisac}{sciezka do pliku .txt, w ktorym chcemy zapisac nasza ramke danych}
}
\value{
Zwracana jest ramka danych o kolumnach z ramek wejsciowych oraz dodatkowej kolumnie \code{tytul_uproszczony}. Kolumny \code{tytul_uproszczony} i \code{rok_produkcji} stanowia klucz glowny w tej ramce danych.
}
\description{
Funkcja \code{polacz_tabele_z_filmami()} laczy wszystkie tabele po kluczu: \code{tytul} i \code{rok_produkcji} w jedna.
}

