% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uprosc_tytul.R
\name{uprosc_tytul}
\alias{uprosc_tytul}
\title{Upraszcza tytuly filmow}
\usage{
uprosc_tytul(wektor_tytulow)
}
\arguments{
\item{wektor_tytulow}{wektor napisow}
}
\value{
wektor napisow
}
\description{
Funkcja \code{uprosc_tytul()} zwraca wektor tytulow pozbawiony spacji, myslnikow, przecinkow, itd., wszystkie litery sa zamienione na male, a znaki z UTF-8 na znaki z alfabetu lacinskiego.
}

