% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wikipedia_charakterystyki.R
\name{wikipedia_charakterystyki}
\alias{wikipedia_charakterystyki}
\title{Pobierz charakterysyki filmow z Wikipedii.}
\usage{
wikipedia_charakterystyki(filmy_wiki, sciezka_wyjscie)
}
\arguments{
\item{filmy_wiki}{Ramki danych o kolumnach z nazwami "tytul_wikipedia", "link_wikipedia", "rok_produkcji". Taka ramke danych mozna otrzymac korzystajac z funkcji \code{wikipedia_tytul_rok_link}.}

\item{sciezka_wyjscie}{Sciezka do pliku tekstowego, w ktorym chcemy zapisac pobrane dane.}
}
\value{
Funkcja \code{wikipedia_charakterystyki} zwraca ramke danych o 10 kolumnach zawierajaca informacje o danym filmie znajdujace sie na Wikipedii.
}
\description{
Funkcja \code{wikipedia_charakterystyki} szuka informacji o filmach na Wikipedii.
}
\details{
Kolumna "link_wikipedia" argumentu \code{filmy_wiki} powinna zawierac linki do filmow z Wikipedii. Wsrod informacji otrzymanych przy pomocy funkcji \code{wikipedia_charakterystyki} znajdujemy: "tytul_wikipedia", "link_wikipedia", "gatunek", "rok_produkcji", "kraj_produkcji", "jezyk", "czas_trwania", "rezyseria", "budzet", "waluta".
}

