% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wikipedia_charakterystyki_nagrody.R
\name{wikipedia_charakterystyki_nagrody}
\alias{wikipedia_charakterystyki_nagrody}
\title{Znajdz tytuly filmow nagrodzone nagrodami na Wikipedii.}
\usage{
wikipedia_charakterystyki_nagrody(nagroda = c("oscar", "zloty glob"),
  sciezka_wyjscie)
}
\arguments{
\item{nagroda}{Wektor napisow zawierajacy slowa "oscar" lub "zloty glob". Domyslnie \code{nagrody = c("oscar", "zloty glob")}. Slowa inne niz podane sa ignorowane.}

\item{sciezka_wyjscie}{Sciezka do pliku tekstowego, w ktorym chcemy zapisac pobrane dane.}
}
\value{
Funkcja \code{wikipedia_charakterystyki_nagrody} zwraca ramke danych o 2 kolumnach zawierajaca tytul filmu i nagrode jaka ten film uzyskal.
}
\description{
Funkcja \code{wikipedia_charakterystyki_nagrody} szuka tytulow filmow na Wikipedii, ktore dostaly Oscara i(lub) Zlotego Globa.
}
\details{
Dane na temat nagrodzonych filmow pochodza z Wikipedii.
}

