% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wikipedia_charakterystyki_obsada.R
\name{wikipedia_charakterystyki_obsada}
\alias{wikipedia_charakterystyki_obsada}
\title{Znajdz obsade filmow na Wikipedii.}
\usage{
wikipedia_charakterystyki_obsada(filmy_wiki, sciezka_wyjscie)
}
\arguments{
\item{filmy_wiki}{Ramka danych o kolumnach z nazwami "tytul_wikipedia", "link_wikipedia", "rok_produkcji". Taka ramke danych mozna otrzymac korzystajac z funkcji \code{wikipedia_tytul_rok_link}.}

\item{sciezka_wyjscie}{Sciezka do pliku tekstowego, w ktorym chcemy zapisac pobrane dane.}
}
\value{
Funkcja \code{wikipedia_charakterystyki_obsada} zwraca ramke danych o 3 kolumnach zawierajaca informacje o obsadzie danych filmow znajdujacych sie na Wiki.
}
\description{
Funkcja \code{wikipedia_charakterystyki_obsada} szuka informacji o obsadzie filmow na Wikipedii.
}
\details{
Kolumna "link_wikipedia" argumentu \code{filmy_wiki} powinna zawierac linki do filmow z Wikipedii.
}

