% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wikipedia_tytul_rok_link.R
\name{wikipedia_tytul_rok_link}
\alias{wikipedia_tytul_rok_link}
\title{Pobierz liste filmow z portalu Wikipedia.}
\usage{
wikipedia_tytul_rok_link(lata, sciezka_wyjscie)
}
\arguments{
\item{lata}{Wektor liczb calkowitych. Funkcja pobierze filmy z Wikipedii ze wszystkich lat podanych w wektorze \code{lata}.}

\item{sciezka_wyjscie}{Sciezka do pliku tekstowego, w ktorym chcemy zapisac pobrane dane.}
}
\value{
Funkcja \code{wikipedia_tytul_rok_link} zwraca ramke danych o trzech kolumnach. Pierwsza kolumna to wektor napisow o nazwie \code{tytul_wikipedia} zawierajaca tytyly filmow znalezionych przez funkcje, druga kolumna to wektor numeryczny \code{rok_produkcji} zawierajaca rok produkcji pobranych filmow, zas trzecia kolumna to wektor napisow o nazwie \code{link_wikipedia}, ktora przechowuje linki do strony danego filmu na Wikipedii.
}
\description{
Funkcja \code{wikipedia_tytul_rok_link} szuka filmow z danego roku (lat) na Wikipedii.
}
\details{
Filmy na Wikipedii sa skatalogowane od 1887 roku. Dlatego sensownie jest za argument lata podac wektor z liczbami calkowitymi wiekszymi badz rownymi 1887.
}

